Enables reverse-beta sorting functionality
Adds the ability for the worker queue to sort by reverse-beta values, so larger values will be intergated first. Since intergration times tend to go with reverse-beta (so the closer beta is to 1, then the longer the intergration time) this makes the whole process more efficient as it no loner allows cores to idle while they wait for more than one core to finish a long intergation. Gains are especially noticable when using large core counts (>= 8 usually), or when the model has a wide range of intergration time (say varying from 2 mins to upwards of 5 hrs).