Updated new task scheduler
This is an update to my existing new reverse-beta task scheduler for three-point function execution. This commit updates it to both a) actually use the argument_cache to store the command line variable, and b) is compatible with beta values that are less than 1/3. This is needed, as previouly a beta=0 task would be intergrated last, however in fact this would take about the same time as a beta=1 task, and so my task scheduler has been updated to cope with these lower beta values.

This commit also fixes a large number of indentation issues in the worker_scheduler.h file.